requirements.txt
# requirements.txt
Flask
google-generativeai
python-dotenv
Flask-Cors
numpy
scipy
gunicorn
gspread
google-auth-oauthlib